Remove logging
I've figured out that the peer connection does not mutate the local description, instead it replaces it, so broadcasting is holding onto the original one which never sees candidates placed into it. We need to instead store the peer connection and read the up to date local description from it.